Backend/requirements.txt
fastapi
uvicorn
pydantic
bcrypt
PyJWT
python-dotenv
langchain
langchain-google-genai
langchain-core
google-generativeai
